Architectural change, gruntfile bugfix
The bugfix is a bit hairy, workingDir.join() sometimes returns a path with
commas, so I added a .replace(',', '/'), but the real issue of why the are
in the path should be addressed. For now, this works.

Also took out AJSEngine.coffee, made AJS.coffee the top-level dependency
which includes actors. AJS should provide helper functions for ad
development, so an engine is redundant (provided by the platform).

Besides that, general dependency updates